+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Generating `variables.tf` for capability modules.
   * Generating `capabilities.tf`, `capabilities.tf.tmpl`, `outputs.tf` for app modules.
 * Fixed `--connection` flags in `blocks new` command.
+* Updated `workspaces select` command to generate `capabilities.tf` from `capabilities.tf.tmpl`.
 
 # 0.0.46 (Feb 24, 2022)
 * Updated CLI to utilize new stack-based API endpoints in Nullstone API.